기본총 데미지증가 적용

+++ Source/TPSProject/public/PlayerFire.h
@@ -25,7 +25,7 @@ public:
 
 
 	UPROPERTY(Editanywhere,BlueprintReadWrite, category=Damage)
-	int sniper_Damage =10;
+	float sniper_Damage =10;
